ragflow-0.22.0/external/langchain-graph/requirements.txt
pydantic==2.10.6
tiktoken==0.9.0
aiohttp==3.11.14
langchain_openai==0.3.9
langgraph==0.3.18
sse-starlette==2.2.1
fastapi==0.115.12
python-multipart==0.0.20
uvicorn==0.34.0
gradio==5.31.0
pydantic_settings==2.8.1
python-jose[cryptography]
passlib[bcrypt]
gunicorn==23.0.0
beautifulsoup4==4.13.4
fake-useragent==2.2.0
